Added new files that includes html and css.
Added files to repo that includes html/CSS and images alongside the site is the @import files that was added also. The CSS for the @import styles are linked to the "page.css" file I am using the first commit to test and see if the imports are working are desired.